+++ apps/contentful-apps/components/content-import/utils.ts
@@ -1,3 +1,6 @@
+import { NodeHtmlMarkdown } from 'node-html-markdown'
+import { richTextFromMarkdown } from '@contentful/rich-text-from-markdown'
+
 import { ReferenceFieldMappingProps } from './ReferenceFieldMapping'
 
 export type FileData = string[][]
@@ -18,4 +21,19 @@ export const extractContentType = (
     validations = field.contentfulField.data.items?.validations ?? []
   }
   return validations.find((v) => v?.linkContentType)?.linkContentType?.[0] ?? ''
+}
+
+export const convertHtmlToContentfulRichText = async (html: string) => {
+  const markdown = NodeHtmlMarkdown.translate(html || '')
+  return richTextFromMarkdown(markdown)
+}
+
+export const downloadCsv = (csvContent: string, filename: string) => {
+  const blob = new Blob([csvContent], { type: 'text/csv' })
+  const url = URL.createObjectURL(blob)
+  const a = document.createElement('a')
+  a.href = url
+  a.download = filename
+  a.click()
+  URL.revokeObjectURL(url)
 }
